# docs/api/gateway-openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "PQC Migration Gateway API",
    "version": "0.1.0",
    "description": "Gateway proxy API for discovery, risk scoring, proof generation, and QASM examples."
  },
  "servers": [
    {
      "url": "http://localhost:8080"
    }
  ],
  "paths": {
...
    "/api/v1/discovery": {
      "post": {
        "summary": "Run discovery scan",
        "requestBody": {
          "required": false,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "address": {
                    "type": "string"
                  },
                  "port": {
                    "type": "integer",
                    "minimum": 1
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Discovery response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "target",
                    "findings"
                  ],
                  "properties": {
                    "target": {
                      "$ref": "#/components/schemas/DiscoveryTarget"
                    },
                    "findings": {
                      "type": "array",
                      "items": { "$ref": "#/components/schemas/Finding" }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/assets": {
      "get": {
        "summary": "List deduplicated discovered assets",
        "responses": {
          "200": {
            "description": "Asset list",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "count",
                    "assets"
                  ],
                  "properties": {
                    "count": {
                      "type": "integer"
                    },
                    "assets": { "type": "array", "items": { "$ref": "#/components/schemas/Asset" } }
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create or upsert a manually supplied asset",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/AssetCreateRequest" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Existing asset upserted",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/AssetCreateResponse" } } }
          },
          "201": {
            "description": "New asset created",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/AssetCreateResponse" } } }
          },
          "400": { "description": "Invalid asset fields" }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "DiscoveryTarget": {
        "type": "object",
        "required": ["address", "port"],
        "properties": {
          "address": { "type": "string" },
          "port": { "type": "integer", "minimum": 1, "maximum": 65535 }
        }
      },
      "Finding": {
        "type": "object",
        "required": ["target", "protocol", "severity", "summary"],
        "properties": {
          "target": { "type": "string" },
          "protocol": { "type": "string" },
          "severity": { "type": "string", "enum": ["critical", "high", "medium", "low", "info", "unknown"] },
          "summary": { "type": "string" }
        }
      },
...
      "Asset": {
        "type": "object",
        "required": ["fingerprint", "target", "protocol", "severity", "summary", "first_seen_at", "last_seen_at", "seen_count"],
        "properties": {
          "fingerprint": { "type": "string" },
          "target": { "type": "string" },
          "protocol": { "type": "string" },
          "severity": { "type": "string", "enum": ["critical", "high", "medium", "low", "info"] },
          "summary": { "type": "string" },
          "first_seen_at": { "type": "string", "format": "date-time" },
          "last_seen_at": { "type": "string", "format": "date-time" },
          "seen_count": { "type": "integer", "minimum": 1 }
        }
      },
      "AssetCreateRequest": {
        "type": "object",
        "required": ["target", "protocol", "severity", "summary"],
        "properties": {
          "target": { "type": "string", "maxLength": 512 },
          "protocol": { "type": "string", "maxLength": 64 },
          "severity": { "type": "string", "enum": ["critical", "high", "medium", "low", "info"] },
          "summary": { "type": "string", "maxLength": 2048 }
        }
      },
      "AssetCreateResponse": {
        "type": "object",
        "required": ["asset", "created"],
        "properties": {
          "asset": { "$ref": "#/components/schemas/Asset" },
          "created": { "type": "boolean" }
        }
      },
...
    }
  }
}